third_party/blink/web_tests/fast/inspector-support/uncaught-dom1-exception-expected.txt

CONSOLE ERROR: Uncaught IndexSizeError: Failed to execute 'splitText' on 'Text': The offset 10000 is larger than the Text node's length.
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
